core: Improve AtomicFile implementation
Fixed issue that system.xml is lost caused by
AtomicFille in shut-down phase with full storage.

+++ core/java/android/util/AtomicFile.java
@@ -149,10 +149,18 @@ public class AtomicFile {
             FileUtils.sync(str);
             try {
                 str.close();
-                mBaseName.delete();
-                mBackupName.renameTo(mBaseName);
+                if (mBackupName.exists()) {
+                    mBaseName.delete();
+                    mBackupName.renameTo(mBaseName);
+                }
             } catch (IOException e) {
                 Log.w("AtomicFile", "failWrite: Got exception:", e);
+            }
+        } else {
+            if (!mBaseName.exists()) {
+                if (mBackupName.exists()) {
+                    mBackupName.renameTo(mBaseName);
+                }
             }
         }
     }
